requirements.txt
# Client to RabbitMQ
pika

# Client to ActiveMQ
stomp.py